Add test steps for MariaDB into the Makefile
There was no test step about Mariadb in the Makefile. 2 have been added
to mimic the tests performed on MySQL.
Moreover, the docker-compose file has been altered in order to add the
MariaDB image. As MariaDB and MySQL are similar, the listening port for
the first one has been moved from 3306 to 6033 to prevent any conflict.
The connection strings have been modified to take the change into
account.